변경사항 일부 적용, 서버 ,클라이언트 작동 구현중
