+++ src/Client/MainProgram.java
@@ -1,0 +1,19 @@
+package Client;
+
+import java.util.ArrayList;
+
+
+public class MainProgram {
+
+    public static void main(String[] args) {
+
+
+
+
+        String serverAddress = "localhost";
+        int serverPort = 54321;
+
+        NPMainPage nmp = new NPMainPage(serverAddress, serverPort);
+
+    }
+}
