--- src/MainProgram.java
@@ -1,10 +1,0 @@
-public class MainProgram {
-
-    public static void main(String[] args) {
-
-        String name= "기업";
-
-        User user = new User(name);
-        NPMainPage np = new NPMainPage(user);
-    }
-}
