+++ src/Server/GameServer.java
@@ -1,0 +1,4 @@
+package Server;
+
+public class GameServer {
+}
